Remove arrowheads from tiefighter line if it is shorter than 15 pixels !!!
I've wanted this for ages and it's so exciting it finally exists. I am not yet adding it to the tiefighter with buttons because that could require a bit more of a code overhaul since the arrowheads might need to be added or removed based on the new line width. And transitioned at the right time.